Add files via upload
Files used for ECC Hamming simulation. Main is main program to run through all 16 bits of 4 bit system. Simulation is a backend simulation to determine if a logic gate being flipped will effect the circuit or not. ErrorPin is a function to determine which transistor fails and whether or not they disrupt the circuit.